refactor: Docs component
