+++ react/src/components/Docs.module.css
@@ -1,0 +1,15 @@
+.card {
+  text-align: left;
+}
+
+.logo {
+  display: grid;
+  place-content: center;
+  padding-block: 48px;
+  width: 100%;
+  background-color: var(--bfc-base-2);
+}
+
+.logo img {
+  height: 100px;
+}
